+++ src/pci.c
@@ -100,18 +100,18 @@
 static int pci_enum_is_64bit(uint32_t value);
 static int pci_enum_is_mmio(uint32_t value);
 
-static inline uint32_t align_up(uint32_t address, uint32_t alignment) {
-    return (address + alignment - 1) & ~(alignment - 1);
+static inline uint64_t align_up(uint64_t address, uint32_t alignment) {
+    return (address + alignment - 1) & ~(uint64_t)(alignment - 1);
 }
 
 static inline uint32_t align_down(uint32_t address, uint32_t alignment) {
     return address & ~(alignment - 1);
 }
 
-static int pci_align_check_up(uint32_t address, uint32_t alignment,
-                              uint64_t limit, uint32_t *aligned)
+static int pci_align_check_up(uint64_t address, uint32_t alignment,
+                              uint64_t limit, uint64_t *aligned)
 {
-    uint32_t a;
+    uint64_t a;
     a = align_up(address, alignment);
     if (a < address || a >= limit)
         return -1;
@@ -363,17 +363,20 @@ static int pci_enum_is_mmio(uint32_t value)
     return (value & PCI_ENUM_MMIND_MASK) == 0;
 }
 
-static int pci_enum_next_aligned32(uint32_t address, uint32_t *next,
+static int pci_enum_next_aligned32(uint64_t address, uint32_t *next,
                                    uint32_t align, uint64_t limit)
 {
-    uintptr_t addr;
+    uint64_t addr;
 
-    addr = (uintptr_t)address;
+    /* 64-bit on purpose: an exhausted pool leaves the cursor at
+     * 0x100000000, which a 32-bit type (uintptr_t included on 32-bit
+     * targets) would truncate back to 0. */
+    addr = address;
     align = align-1;
-    addr = (addr + align) & (~align);
+    addr = (addr + align) & (~(uint64_t)align);
     if (addr > 0xffffffff)
         return -1;
-    if (addr < (uintptr_t)address)
+    if (addr < address)
         return -1;
     if (addr >= limit)
         return -1;
@@ -421,7 +424,7 @@ static int pci_program_bar(uint8_t bus, uint8_t dev, uint8_t fun,
     uint32_t length, align;
     uint8_t bar_off;
     int is_prefetch;
-    uint32_t *base;
+    uint64_t *base;
     uint64_t limit;
     uint32_t reg;
     int is_mmio;
@@ -524,7 +527,7 @@ static int pci_program_bar(uint8_t bus, uint8_t dev, uint8_t fun,
     pci_config_write32(bus, dev, fun, bar_off, bar_value);
     if (*is_64bit)
         pci_config_write32(bus, dev, fun, bar_off + 4, 0x0);
-    *base = bar_value + length;
+    *base = (uint64_t)bar_value + length;
     PCI_DEBUG_PRINTF("PCI enum: %s bus: %x:%x.%x bar: %d [%x,%x] (0x%x %s %s)\r\n",
                     (is_mmio ? "mm" : "io"), bus, dev, fun, bar_idx, bar_value,
                      bar_value + length, length, (*is_64bit) ? "64bit" : "",
@@ -617,14 +620,14 @@ static inline void pci_dump_bridge(uint8_t bus, uint8_t dev, uint8_t fun)
 static int pci_program_bridge(uint8_t bus, uint8_t dev, uint8_t fun,
                                    struct pci_enum_info *info)
 {
-    uint32_t prefetch_start;
-    uint32_t mem_start;
-    uint32_t io_start;
+    uint64_t prefetch_start;
+    uint64_t mem_start;
+    uint64_t io_start;
     uint32_t orig_cmd;
     uint8_t  saved_bus;
-    uint32_t saved_mem;
-    uint32_t saved_pf;
-    uint32_t saved_io;
+    uint64_t saved_mem;
+    uint64_t saved_pf;
+    uint64_t saved_io;
     int ret;
 
     saved_bus = info->curr_bus_number;
@@ -967,16 +970,17 @@ int pci_enum_do(void)
     ret = pci_enum_bus(0, &enum_info);
 
     PCI_DEBUG_PRINTF("PCI Memory Mapped I/O range [0x%x,0x%x] (0x%x)\r\n",
-                     (uint32_t)PCI_MMIO32_BASE, enum_info.mem,
-                     enum_info.mem - PCI_MMIO32_BASE);
+                     (uint32_t)PCI_MMIO32_BASE, (uint32_t)enum_info.mem,
+                     (uint32_t)(enum_info.mem - PCI_MMIO32_BASE));
 
     PCI_DEBUG_PRINTF("PCI Memory Mapped I/O range (prefetch) [0x%x,0x%x] (0x%x)\r\n",
-                     (uint32_t)PCI_MMIO32_PREFETCH_BASE, enum_info.mem_pf,
-                     enum_info.mem_pf - PCI_MMIO32_PREFETCH_BASE);
+                     (uint32_t)PCI_MMIO32_PREFETCH_BASE,
+                     (uint32_t)enum_info.mem_pf,
+                     (uint32_t)(enum_info.mem_pf - PCI_MMIO32_PREFETCH_BASE));
 
     PCI_DEBUG_PRINTF("PCI I/O range [0x%x,0x%x] (0x%x)\r\n",
-                     (uint32_t)PCI_IO32_BASE, enum_info.io,
-                     enum_info.io - PCI_IO32_BASE);
+                     (uint32_t)PCI_IO32_BASE, (uint32_t)enum_info.io,
+                     (uint32_t)(enum_info.io - PCI_IO32_BASE));
 
     return ret;
 }
